refactor(sidebar): extract library and inspector action hooks (#436)
